[CONJS-336] connection attribute _server_host send host, but IP resulting of name resolution

+++ test/integration/auth-plugin.test.js
@@ -645,15 +645,25 @@ describe.concurrent('authentication plugin', () => {
       return;
     }
 
+    let strictPasswordValidation;
+    try {
+      const res = await shareConn.query('SELECT @@global.strict_password_validation');
+      strictPasswordValidation = res[0]['@@global.strict_password_validation'] === 1;
+    } catch (e) {
+      strictPasswordValidation = false;
+    }
+
     await shareConn.query('drop user verifParsec' + getHostSuffix()).catch(() => {});
     await shareConn.query(
       'CREATE USER verifParsec' + getHostSuffix() + " IDENTIFIED VIA parsec USING PASSWORD('MySup8%rPassw@ord')"
     );
     await shareConn.query('GRANT SELECT on `' + Conf.baseConfig.database + '`.* to verifParsec' + getHostSuffix());
 
-    await shareConn.query('drop user verifParsec2' + getHostSuffix()).catch(() => {});
-    await shareConn.query('CREATE USER verifParsec2' + getHostSuffix() + " IDENTIFIED VIA parsec USING PASSWORD('')");
-    await shareConn.query('GRANT SELECT on `' + Conf.baseConfig.database + '`.* to verifParsec2' + getHostSuffix());
+    if (!strictPasswordValidation) {
+      await shareConn.query('drop user verifParsec2' + getHostSuffix()).catch(() => {});
+      await shareConn.query('CREATE USER verifParsec2' + getHostSuffix() + " IDENTIFIED VIA parsec USING PASSWORD('')");
+      await shareConn.query('GRANT SELECT on `' + Conf.baseConfig.database + '`.* to verifParsec2' + getHostSuffix());
+    }
 
     let conn = await createConnection({
       user: 'verifParsec',
